all requirements fullfilled

+++ routes/web.php
@@ -26,10 +26,10 @@ Route::middleware('auth')->group(function () {
 });
 
 Route::group(['prefix' => 'settings', 'middleware' => ['web', 'auth']], function () {
-    Route::get('/system/users/permissions', [UserPermissionController::class, 'index'])->name('system.permissions.index');
-    Route::post('/system/users/permissions', [UserPermissionController::class, 'store'])->name('system.permissions.store');
-    Route::get('/system/users/permissions/create', [UserPermissionController::class, 'create'])->name('system.permissions.create');
-    Route::delete('/system/users/permissions/{id}', [UserPermissionController::class, 'destroy'])->name('system.permissions.destroy');
+    Route::middleware(['role:superadmin'])->get('/system/users/permissions', [UserPermissionController::class, 'index'])->name('system.permissions.index');
+    Route::middleware(['role:superadmin'])->post('/system/users/permissions', [UserPermissionController::class, 'store'])->name('system.permissions.store');
+    Route::middleware(['role:superadmin'])->get('/system/users/permissions/create', [UserPermissionController::class, 'create'])->name('system.permissions.create');
+    Route::middleware(['role:superadmin'])->delete('/system/users/permissions/{id}', [UserPermissionController::class, 'destroy'])->name('system.permissions.destroy');
 });
 
 require __DIR__ . '/auth.php';
